Add features and optimize
Add totvar, quadvar, and fslope features.

Optimize several of the longer running feature functions as well as the
handling of the new feature columns in the data frames in order to
reduce the runtime of the script.

Also standardize the light curve times and magnitudes to use (m,1) numpy
arrays to keep it consistent throughout the script.