Set script to take cli arguments for settings
Set the script so that it takes in settings like the number of rows to
process and whether to save the intermediate curve files as command line
arguments.

This allows these settings to be easliy changed without having to alter
the source code.

Also remove the filling of nan values for extracted features in the case
of a missing light curve, as this is already done automatically by
Pandas.