Add config flags and implement period
Add some configuration flags to specify the columns for the star id and
period as they are required for the feature extraction to work and they
may vary from data file to data file.

Also change the curve file format to fit a simple (time, magnitude, error)
structure rather than previous inclusion of an unused star id column.
This will make the feature extraction more generally applicable.

Also create a function to calculate the period of a light curve using
astropy's Lomb Scargle periodogram, but leave it commented out as it
takes a long time to run and greatly slows down the feature extraction.