Change accuracy to F1 score and misc changes
Change the classifyier scoring to use F1 score rather than accuracy, as
it will be a better measure of their performance.

Print out a list of the most useful features based on how many times
they occur in the best features for cells in the matrix.

Set the html table to print in a format that works better for styling.