Implement cross validation and grid writing
Set the feature importances and classifier scoring to use cross
validation.

Also set the script to write the html feature matrix to a file when it
is run, so that this does not have to be done manually.

Also change classifier scoring to display binary confusion matrices
instead of scores.