Change the scoring from F1 to MCC
Change the scoring function used for evaluating the classifiers to be a
normalized MCC rather than F1, as it is more useful in this use case.

Also change the confusion matrices to be the ones from the first
cross validation rather than a mean, as taking the mean does not make
sense.